requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dj-database-url==0.5.0
Django==4.0.6
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-livereload==1.7
django-livereload-server==0.4
django-on-heroku==1.1.2
django-storages==1.12.3
djangorestframework==3.13.1
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.0
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
urllib3==1.26.9
Werkzeug==2.1.2
whitenoise==6.2.0
zipp==3.8.0
